+++ engine/net/src/actors/admin/mod.rs
@@ -15,3 +15,6 @@ pub use list_invitations::*;
 
 pub mod create_user;
 pub use create_user::*;
+
+pub mod usage_stats;
+pub use usage_stats::*;