+++ engine/net/src/broker.rs
@@ -635,11 +635,13 @@ impl Broker {
         mut connection: ConnectionBase,
         remote_bind_address: BindAddress,
         local_bind_address: BindAddress,
+        net_stats: Arc<std::sync::Mutex<NetStats>>,
     ) -> Result<(), NetError> {
         if self.closing {
             return Err(NetError::Closing);
         }
 
+        let fsm = connection.fsm.as_ref().map(Arc::clone);
         let join: mpsc::UnboundedReceiver<Either<NetError, X25519PrivKey>> =
             connection.take_shutdown();
         if self
@@ -658,6 +660,9 @@ impl Broker {
             mut join: Receiver<Either<NetError, X25519PrivKey>>,
             remote_bind_address: BindAddress,
             local_bind_address: BindAddress,
+            fsm: Option<Arc<Mutex<NoiseFSM>>>,
+            // The critical section is tiny so sync Mutex is ok.
+            net_stats: Arc<std::sync::Mutex<NetStats>>,
         ) -> ResultSend<()> {
             async move {
                 let res = join.next().await;
@@ -690,12 +695,39 @@ impl Broker {
                             .remove_anonymous(remote_bind_address, local_bind_address);
                     }
                 }
+                if let Some(fsm) = fsm {
+                    let fsm = fsm.lock().await;
+                    match fsm.user_id() {
+                        Ok(user_id) => match BROKER.write().await.get_server_broker() {
+                            Err(e) => log_err!(
+                                "Cannot get the server broker, won't \
+                                              save net stats: {}",
+                                e
+                            ),
+                            Ok(sb) => {
+                                let sb = sb.read().await;
+                                let net_stats = net_stats.lock().unwrap();
+                                sb.usage_stats_update_for_user(&user_id, &net_stats);
+                            }
+                        },
+                        Err(err) => {
+                            log_debug!("No user_id to account usage stats to (instead: {})", err);
+                        }
+                    }
+                } else {
+                    log_debug!("NET STATS: No FSM");
+                }
             }
             .await;
             Ok(())
         }
-        spawn_and_log_error(watch_close(join, remote_bind_address, local_bind_address));
-
+        spawn_and_log_error(watch_close(
+            join,
+            remote_bind_address,
+            local_bind_address,
+            fsm,
+            net_stats,
+        ));
         Ok(())
     }
 
@@ -879,6 +911,7 @@ impl Broker {
         cnx.probe(ip, port).await
     }
 
+    // Send an admin command to remote_peer_id
     pub async fn admin<
         A: Into<ProtocolMessage>
             + Into<AdminRequestContentV0>
@@ -889,9 +922,9 @@ impl Broker {
     >(
         &mut self,
         cnx: Box<dyn IConnect>,
-        peer_privk: PrivKey,
-        peer_pubk: PubKey,
-        remote_peer_id: DirectPeerId,
+        peer_privk: PrivKey,          // This broker's privk
+        peer_pubk: PubKey,            // This broker's pubk
+        remote_peer_id: DirectPeerId, // Remote broker's pubk
         user: PubKey,
         user_priv: PrivKey,
         addr: BindAddress,